Bugfix & Active Directory LDAP compatibility
Line 71:  Bugfix: defined requires args to be in quotes
Line 72:  Compat: This works for Active Directory, but might horribly break other directory services.  Creating a constant in preferences called "ACTIVE_DIRECTORY" and checking that and determining whether to pass $username or $result[0]['dn'] would be better.
Line 100:  Compat: memberUid is meaningless in AD.  So, allow the user to specify their own LDAP_GROUP_ATTRIBUTE.  There is a corresponding change in settings.